requirements.txt
diffusers
transformers
accelerate
pillow